refactor(doctor): extract empty allowlist policy warnings
