+++ src/commands/doctor-config-flow.ts
@@ -56,6 +56,7 @@ import {
   resolveConfigPathTarget,
   stripUnknownConfigKeys,
 } from "./doctor-config-analysis.js";
+import { detectEmptyAllowlistPolicy } from "./doctor-empty-allowlist-warnings.js";
 import { normalizeCompatibilityConfigValues } from "./doctor-legacy-config.js";
 import type { DoctorOptions } from "./doctor-prompter.js";
 import { autoMigrateLegacyStateDir } from "./doctor-state-migrations.js";
@@ -1174,138 +1175,6 @@ async function maybeRepairAllowlistPolicyAllowFrom(cfg: OpenClawConfig): Promise
     return { config: cfg, changes: [] };
   }
   return { config: next, changes };
-}
-
-/**
- * Scan all channel configs for dmPolicy="allowlist" without any allowFrom entries.
- * This configuration blocks all DMs because no sender can match the empty
- * allowlist. Common after upgrades that remove external allowlist
- * file support.
- */
-function detectEmptyAllowlistPolicy(cfg: OpenClawConfig): string[] {
-  const channels = cfg.channels;
-  if (!channels || typeof channels !== "object") {
-    return [];
-  }
-
-  const warnings: string[] = [];
-
-  const usesSenderBasedGroupAllowlist = (channelName?: string): boolean => {
-    if (!channelName) {
-      return true;
-    }
-    // These channels enforce group access via channel/space config, not sender-based
-    // groupAllowFrom lists.
-    return !(channelName === "discord" || channelName === "slack" || channelName === "googlechat");
-  };
-
-  const allowsGroupAllowFromFallback = (channelName?: string): boolean => {
-    if (!channelName) {
-      return true;
-    }
-    // Keep doctor warnings aligned with runtime access semantics.
-    return !(
-      channelName === "googlechat" ||
-      channelName === "imessage" ||
-      channelName === "matrix" ||
-      channelName === "msteams" ||
-      channelName === "irc"
-    );
-  };
-
-  const checkAccount = (
-    account: Record<string, unknown>,
-    prefix: string,
-    parent?: Record<string, unknown>,
-    channelName?: string,
-  ) => {
-    const dmEntry = account.dm;
-    const dm =
-      dmEntry && typeof dmEntry === "object" && !Array.isArray(dmEntry)
-        ? (dmEntry as Record<string, unknown>)
-        : undefined;
-    const parentDmEntry = parent?.dm;
-    const parentDm =
-      parentDmEntry && typeof parentDmEntry === "object" && !Array.isArray(parentDmEntry)
-        ? (parentDmEntry as Record<string, unknown>)
-        : undefined;
-    const dmPolicy =
-      (account.dmPolicy as string | undefined) ??
-      (dm?.policy as string | undefined) ??
-      (parent?.dmPolicy as string | undefined) ??
-      (parentDm?.policy as string | undefined) ??
-      undefined;
-
-    const topAllowFrom =
-      (account.allowFrom as Array<string | number> | undefined) ??
-      (parent?.allowFrom as Array<string | number> | undefined);
-    const nestedAllowFrom = dm?.allowFrom as Array<string | number> | undefined;
-    const parentNestedAllowFrom = parentDm?.allowFrom as Array<string | number> | undefined;
-    const effectiveAllowFrom = topAllowFrom ?? nestedAllowFrom ?? parentNestedAllowFrom;
-
-    if (dmPolicy === "allowlist" && !hasAllowFromEntries(effectiveAllowFrom)) {
-      warnings.push(
-        `- ${prefix}.dmPolicy is "allowlist" but allowFrom is empty — all DMs will be blocked. Add sender IDs to ${prefix}.allowFrom, or run "${formatCliCommand("openclaw doctor --fix")}" to auto-migrate from pairing store when entries exist.`,
-      );
-    }
-
-    const groupPolicy =
-      (account.groupPolicy as string | undefined) ??
-      (parent?.groupPolicy as string | undefined) ??
-      undefined;
-
-    if (groupPolicy === "allowlist" && usesSenderBasedGroupAllowlist(channelName)) {
-      const rawGroupAllowFrom =
-        (account.groupAllowFrom as Array<string | number> | undefined) ??
-        (parent?.groupAllowFrom as Array<string | number> | undefined);
-      // Match runtime semantics: resolveGroupAllowFromSources treats
-      // empty arrays as unset and falls back to allowFrom.
-      const groupAllowFrom = hasAllowFromEntries(rawGroupAllowFrom) ? rawGroupAllowFrom : undefined;
-      const fallbackToAllowFrom = allowsGroupAllowFromFallback(channelName);
-      const effectiveGroupAllowFrom =
-        groupAllowFrom ?? (fallbackToAllowFrom ? effectiveAllowFrom : undefined);
-
-      if (!hasAllowFromEntries(effectiveGroupAllowFrom)) {
-        if (fallbackToAllowFrom) {
-          warnings.push(
-            `- ${prefix}.groupPolicy is "allowlist" but groupAllowFrom (and allowFrom) is empty — all group messages will be silently dropped. Add sender IDs to ${prefix}.groupAllowFrom or ${prefix}.allowFrom, or set groupPolicy to "open".`,
-          );
-        } else {
-          warnings.push(
-            `- ${prefix}.groupPolicy is "allowlist" but groupAllowFrom is empty — this channel does not fall back to allowFrom, so all group messages will be silently dropped. Add sender IDs to ${prefix}.groupAllowFrom, or set groupPolicy to "open".`,
-          );
-        }
-      }
-    }
-  };
-
-  for (const [channelName, channelConfig] of Object.entries(
-    channels as Record<string, Record<string, unknown>>,
-  )) {
-    if (!channelConfig || typeof channelConfig !== "object") {
-      continue;
-    }
-    checkAccount(channelConfig, `channels.${channelName}`, undefined, channelName);
-
-    const accounts = channelConfig.accounts;
-    if (accounts && typeof accounts === "object") {
-      for (const [accountId, account] of Object.entries(
-        accounts as Record<string, Record<string, unknown>>,
-      )) {
-        if (!account || typeof account !== "object") {
-          continue;
-        }
-        checkAccount(
-          account,
-          `channels.${channelName}.accounts.${accountId}`,
-          channelConfig,
-          channelName,
-        );
-      }
-    }
-  }
-
-  return warnings;
 }
 
 type ExecSafeBinCoverageHit = {
